requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
httpx==0.25.2
openai==1.3.7
python-dotenv==1.0.0
azure-identity==1.15.0
azure-core==1.35.0
msal==1.33.0
msal-extensions==1.3.1
cryptography>=2.5
requests>=2.21.0 